[c++] Fix potential loop overflows

+++ cpp/inc/bond/protocol/compact_binary.h
@@ -473,7 +473,7 @@ protected:
         uint32_t                                size;
 
         ReadContainerBegin(size, element_type);
-        for (uint32_t i = 0; i < size; ++i)
+        for (int64_t i = 0; i < size; ++i)
         {
             SkipType(element_type.first);
             SkipType(element_type.second);
@@ -535,14 +535,14 @@ protected:
     {
         if (v2 == _version)
         {
-            for (uint32_t i = 0; i < size; ++i)
+            for (int64_t i = 0; i < size; ++i)
             {
                 SkipStructV2();
             }
         }
         else
         {
-            for (uint32_t i = 0; i < size; ++i)
+            for (int64_t i = 0; i < size; ++i)
             {
                 SkipStructV1();
             }
@@ -556,7 +556,7 @@ protected:
                                 || T == BT_SET || T == BT_LIST || T == BT_MAP)>::type
     SkipType(uint32_t size)
     {
-        for (uint32_t i = 0; i < size; ++i)
+        for (int64_t i = 0; i < size; ++i)
         {
             SkipType<T>();
         }
